% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_lifecycle_to_package_doc}
\alias{add_lifecycle_to_package_doc}
\title{Add {lifecycle} components to package doc file}
\usage{
add_lifecycle_to_package_doc()
}
\value{

}
\description{
Add {lifecycle} components to package doc file
}
