% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask user preferences}
\usage{
ask(minimal = FALSE)
}
\arguments{
\item{minimal}{\link{logical(1)} Reduce the questions to just the essentials}
}
\value{

}
\description{
Ask user preferences
}
