% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{block_append}
\alias{block_append}
\title{Append block}
\usage{
block_append(lines, value, block_bounds, only_if_new = TRUE, strict = TRUE)
}
\arguments{
\item{strict}{}
}
\value{

}
\description{
Append block
}
