% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{block_create}
\alias{block_create}
\title{Create block}
\usage{
block_create(
  value,
  block_start,
  block_end,
  block_prefix = character(),
  block_suffix = character()
)
}
\arguments{
\item{block_suffix}{}
}
\value{

}
\description{
Create block
}
