% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{block_find}
\alias{block_find}
\title{Find block between start and end identifiers}
\usage{
block_find(lines, block_start, block_end)
}
\arguments{
\item{block_end}{}
}
\value{

}
\description{
Find block between start and end identifiers
}
