% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_local_package}
\alias{create_local_package}
\title{Create a local package}
\usage{
create_local_package(
  dir = fs::file_temp(pattern = "testpkg"),
  env = parent.frame()
)
}
\arguments{
\item{env}{}
}
\value{

}
\description{
Mainly created for clean testing as described in Jenny Bryan's
\href{https://www.tidyverse.org/blog/2020/04/self-cleaning-test-fixtures/}{Self-cleaning test fixtures}
and to facilitate examples that don't mess things up for the package users.
}
