% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_sandbox_package}
\alias{create_sandbox_package}
\title{Create a sandbox package}
\usage{
create_sandbox_package(
  dir = fs::file_temp(pattern = "testpkg"),
  env = parent.frame(),
  defer = TRUE
)
}
\arguments{
\item{env}{}
}
\value{

}
\description{
Mainly created for clean testing as described in Jenny Bryan's
\href{https://www.tidyverse.org/blog/2020/04/self-cleaning-test-fixtures/}{Self-cleaning test fixtures}
and to facilitate examples that don't mess things up for the package users.
}
