% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_ci}
\alias{ensure_ci}
\title{Ensure Continuous Integration}
\usage{
ensure_ci(
  platform = valid_ci_platforms("github_actions"),
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\value{
\link{logical(1)}
}
\description{
Ensure Continuous Integration
}
