% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_ci_test_coverage}
\alias{ensure_ci_test_coverage}
\title{Ensure Continuous Integration test coverage}
\usage{
ensure_ci_test_coverage(
  service = valid_ci_test_coverage_services("codecov"),
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE,
  info = FALSE,
  .trace = FALSE
)
}
\arguments{
\item{strict}{}
}
\value{
\link{logical(1)}
}
\description{
Ensure Continuous Integration test coverage
}
