% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_ci_test_coverage_link}
\alias{ensure_ci_test_coverage_link}
\title{Ensure CI test coverage service is linked to CI platform}
\usage{
ensure_ci_test_coverage_link(
  ci_test_coverage_service = valid_ci_test_coverage_services("codecov"),
  ci_platform = valid_ci_platforms("github_actions"),
  info = FALSE,
  .trace = FALSE
)
}
\arguments{
\item{ci_platform}{}
}
\value{

}
\description{
Ensure CI test coverage service is linked to CI platform
}
