% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_dependency_management}
\alias{ensure_dependency_management}
\title{Ensure dependency management}
\usage{
ensure_dependency_management(
  package = valid_dep_management_packages("renv"),
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\value{
\link{logical(1)}
}
\description{
Ensure dependency management
}
