% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_generic}
\alias{ensure_generic}
\title{Generic function for ensuring things that \code{{goodstart}} can handle}
\usage{
ensure_generic(
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  fn,
  fn_dep = NULL,
  deps = character(),
  deps_type = "Suggests",
  path = character(),
  requires_restart = FALSE
)
}
\arguments{
\item{requires_restart}{}
}
\description{
Generic function for ensuring things that \code{{goodstart}} can handle
}
