% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_readme_rmd}
\alias{ensure_readme_rmd}
\title{Ensure README.Rmd exists and is populated based on project data}
\usage{
ensure_readme_rmd(
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  open = rlang::is_interactive(),
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\value{
\code{\link[=logical]{logical()}}
}
\description{
Ensure README.Rmd exists and is populated based on project data
}
