% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_removed_namespace_default}
\alias{ensure_removed_namespace_default}
\title{Ensure that default NAMESPACE is removed}
\usage{
ensure_removed_namespace_default(info = FALSE, .trace = FALSE)
}
\value{
\link{logical(1)}
}
\description{
When \code{NAMESPACE} contains \verb{exportPattern(\\"^[[:alpha:]]+\\")} it is removed to
allow \code{{roxygen2}} to control \code{NAMESPACE } content.
}
