% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_renv_active}
\alias{ensure_renv_active}
\title{Ensure \code{{renv}} project is activated}
\usage{
ensure_renv_active(ensure_deps = TRUE, add_to_suggested = TRUE, strict = TRUE)
}
\arguments{
\item{strict}{}
}
\description{
Ensure \code{{renv}} project is activated
}
