% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_renv_upgraded}
\alias{ensure_renv_upgraded}
\title{Ensure \code{{renv}} project is upgraded}
\usage{
ensure_renv_upgraded(
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\description{
Ensure \code{{renv}} project is upgraded
}
