% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_roxygen}
\alias{ensure_roxygen}
\title{Ensure Roxygen documentation via \code{{roxygen2}} is enabled}
\usage{
ensure_roxygen(
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  info = FALSE,
  .trace = FALSE
)
}
\arguments{
\item{info}{}
}
\description{
Ensure Roxygen documentation via \code{{roxygen2}} is enabled
}
