% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_roxygen_namespace}
\alias{ensure_roxygen_namespace}
\title{#' @importFrom roxygen2md roxygen2md
ensure_roxygen2md <- function(
install_deps_if_missing = TRUE,
add_to_description_suggests = TRUE
) {
ensure_generic(
install_deps_if_missing = install_deps_if_missing,
add_to_description_suggests = add_to_description_suggests,
fn = roxygen2md::roxygen2md,
deps = "roxygen2md"
)
}}
\usage{
ensure_roxygen_namespace(install_deps_if_missing = TRUE)
}
\description{
#' @importFrom roxygen2md roxygen2md
ensure_roxygen2md <- function(
install_deps_if_missing = TRUE,
add_to_description_suggests = TRUE
) {
ensure_generic(
install_deps_if_missing = install_deps_if_missing,
add_to_description_suggests = add_to_description_suggests,
fn = roxygen2md::roxygen2md,
deps = "roxygen2md"
)
}
}
