% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_unit_testing}
\alias{ensure_unit_testing}
\title{Ensure unit testing package}
\usage{
ensure_unit_testing(
  package = valid_unit_test_packages("testthat"),
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\value{
\link{logical(1)}
}
\description{
Ensure unit testing package
}
