% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_unit_testing_test_coverage}
\alias{ensure_unit_testing_test_coverage}
\title{Ensure test coverage package \code{{covr}}}
\usage{
ensure_unit_testing_test_coverage(
  package = valid_unit_test_coverage_packages("covr"),
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\description{
Ensure test coverage package \code{{covr}}
}
