% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_unit_testing_testthat}
\alias{ensure_unit_testing_testthat}
\title{Ensure \code{{testthat}} infrastructure}
\usage{
ensure_unit_testing_testthat(
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE
)
}
\description{
Ensure \code{{testthat}} infrastructure
}
