% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{envvy__deparse}
\alias{envvy__deparse}
\title{Deparse environment variables}
\usage{
envvy__deparse(env_vars)
}
\arguments{
\item{env_vars}{}
}
\value{
\code{\link[=list]{list()}} List with deparsed environment variable expression as
\link{character(1)} per list element
}
\description{
Deparse environment variables
}
