% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{envvy__parse}
\alias{envvy__parse}
\title{Parse environment variables}
\usage{
envvy__parse(env_vars)
}
\arguments{
\item{env_vars}{}
}
\value{
\code{\link[=list]{list()}} List with parsed values
}
\description{
Parse environment variables
}
