% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{exists_goodstart_config}
\alias{exists_goodstart_config}
\title{Check existence of config file}
\usage{
exists_goodstart_config(path = get_path_config_file())
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Check existence of config file
}
