% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{goodstart_config}
\alias{goodstart_config}
\title{Config for \code{{goodstart}}}
\usage{
goodstart_config(
  use_github_user_name = "rappster",
  use_github_repo_name = get_package_name(),
  use_github_auth = "ssh"
)
}
\arguments{
\item{use_github_auth}{}
}
\value{

}
\description{
Config for \code{{goodstart}}
}
