% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{list_goodstart_configs}
\alias{list_goodstart_configs}
\title{List available config file}
\usage{
list_goodstart_configs(path = path_config_file(id = id), id = character())
}
\arguments{
\item{id}{}
}
\value{

}
\description{
List available config file
}
