% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mimick_rstudio_package}
\alias{mimick_rstudio_package}
\title{Mimick R package project as created from within RStudio}
\usage{
mimick_rstudio_package()
}
\value{
Project path as returned by \code{\link[usethis:proj_utils]{usethis::proj_path()}}
}
\description{
Mimick R package project as created from within RStudio
}
