% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_goodstart_config}
\alias{read_goodstart_config}
\title{Read config for \code{{goodstart}}}
\usage{
read_goodstart_config(path = path_config_file(id = id), id = character())
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Read config for \code{{goodstart}}
}
