% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue.R
\name{skeval__exists_value}
\alias{skeval__exists_value}
\title{Check if hash key exists in key-value store}
\usage{
skeval__exists_value(key)
}
\arguments{
\item{key}{}
}
\value{

}
\description{
Check if hash key exists in key-value store
}
