% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue.R
\name{skeval__read_value}
\alias{skeval__read_value}
\title{Read value from key-value store}
\usage{
skeval__read_value(key)
}
\arguments{
\item{key}{}
}
\value{

}
\description{
Read value from key-value store
}
