% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\title{Use custom README.Rmd template}
\usage{
use_readme_rmd(open = rlang::is_interactive())
}
\arguments{
\item{open}{}
}
\description{
Use custom README.Rmd template
}
