% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_goodstart_config}
\alias{write_goodstart_config}
\title{Write config for \code{{goodstart}}}
\usage{
write_goodstart_config(
  config = goodstart_config(),
  path = path_config_file(id = id),
  id = character()
)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Write config for \code{{goodstart}}
}
