% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_backlog_rmd}
\alias{ensure_backlog_rmd}
\title{Ensure BACKLOG.Rmd}
\usage{
ensure_backlog_rmd(open = rlang::is_interactive())
}
\arguments{
\item{open}{}
}
\value{
\link{logical(1)}
}
\description{
Ensure BACKLOG.Rmd
}
