% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_readme}
\alias{ensure_readme}
\title{Ensure README}
\usage{
ensure_readme(
  type = valid_readme_types("rmd"),
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  open = rlang::is_interactive(),
  strict = TRUE,
  info = FALSE,
  .trace = FALSE
)
}
\arguments{
\item{info}{}
}
\value{
\code{\link[=logical]{logical()}}
}
\description{
Ensure README
}
