% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.R
\name{ensure_roxygen_namespace}
\alias{ensure_roxygen_namespace}
\title{#' @importFrom roxygen2md roxygen2md
ensure_roxygen2md <- function(
ensure_deps = TRUE,
add_to_suggested = TRUE
) {
ensure_generic(
ensure_deps = ensure_deps,
add_to_suggested = add_to_suggested,
fn = roxygen2md::roxygen2md(),
deps = "roxygen2md"
)
}}
\usage{
ensure_roxygen_namespace(
  ensure_deps = TRUE,
  add_to_suggested = TRUE,
  strict = TRUE
)
}
\description{
#' @importFrom roxygen2md roxygen2md
ensure_roxygen2md <- function(
ensure_deps = TRUE,
add_to_suggested = TRUE
) {
ensure_generic(
ensure_deps = ensure_deps,
add_to_suggested = add_to_suggested,
fn = roxygen2md::roxygen2md(),
deps = "roxygen2md"
)
}
}
