% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{exists_goodstart_config}
\alias{exists_goodstart_config}
\title{Check existence of config file}
\usage{
exists_goodstart_config(path = path_config_file(id = id), id = character())
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Check existence of config file
}
