% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_goodstart_config}
\alias{read_goodstart_config}
\title{Read config for \code{{goodstart}}}
\usage{
read_goodstart_config(path = get_path_config_file())
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Read config for \code{{goodstart}}
}
